pyproject.toml
[project]
name = "withgames-discord-bot"
version = "0.1.0"
description = "A feature-rich Discord bot for managing game event recruitment"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Y-Kuro-u" }
]

dependencies = [
    "discord.py>=2.4.0",
    "google-cloud-firestore>=2.16.0",
    "google-auth>=2.29.0",
    "google-cloud-secret-manager>=2.18.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
    "python-dateutil>=2.9.0",
    "click>=8.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
